// src/app/models/rutas.ts
export interface RutaInterface {
    idR?: string;
    fechainicio?: string;
    activo?: boolean;
    payload?: string;

    proyectodsc?: string;
    idproyecto?: string;

    email?: string;
    idusuario?: string;

    idtienda?: string
    canaldsc?: string;
    cadenadsc?: string;
    formatodsc?: string;
    sucursal?: string;
    estadodsc?: string;
    ciudaddsc?: string;
    calle?: string;
    colonia?: string;
    cp?: string;
}